Initial Commit-
A simple google chrome extension that goes out to flicker and grabs pictures.  If you want to change which pictures it grabs, you have to change the url in the javascript file.  Also need to make sure you send the parameter format=XML at the end of the URL.

Future Things to Do:
Make a box where the user can "put" the url, take this add the format=XML tag and use that to grab pictures.